requirements/ico.txt
Pillow==8.1.0
